fix warning of ios version

+++ MDraw/MDrawFreeline.m
@@ -70,8 +70,8 @@
     
     return CGRectMake(tl.x,
                       tl.y,
-                      fabsf(rb.x - tl.x),
-                      fabsf(rb.y - tl.y));
+                      fabs(rb.x - tl.x),
+                      fabs(rb.y - tl.y));
     
     
 }
